Fixed stripe onboarding
